Complete re-write of intermediate interpreter to fix most calculations.
